requirements.txt
atlassian-python-api
py-trello
pystache
pyinstaller